# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Image:
    properties:
      url:
        type: string
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.OrderInBasket:
    properties:
      available_count:
        type: integer
      city_id:
        type: integer
      count:
        type: integer
      delivery:
        type: boolean
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Image'
        type: array
      in_favourites:
        type: boolean
      owner_id:
        type: integer
      price:
        type: integer
      product_id:
        type: integer
      safe_deal:
        type: boolean
      saler_id:
        type: integer
      title:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.ProductInFeed:
    properties:
      available_count:
        type: integer
      city_id:
        type: integer
      delivery:
        type: boolean
      favourites:
        type: integer
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Image'
        type: array
      in_favourites:
        type: boolean
      is_active:
        type: boolean
      price:
        type: integer
      safe_deal:
        type: boolean
      title:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError'
      status:
        type: integer
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBody:
    properties:
      message:
        type: string
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError:
    properties:
      error:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBody'
      status:
        type: integer
    type: object
  internal_models.OrderChanges:
    properties:
      count:
        type: integer
      id:
        type: integer
      status:
        type: integer
    type: object
  internal_models.PreOrder:
    properties:
      count:
        type: integer
      product_id:
        type: integer
    type: object
...
  internal_pkg_product_delivery.OrderListResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.OrderInBasket'
        type: array
      status:
        type: integer
    type: object
  internal_pkg_product_delivery.OrderResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.OrderInBasket'
      status:
        type: integer
    type: object
  internal_pkg_product_delivery.ProductListResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.ProductInFeed'
        type: array
      status:
        type: integer
    type: object
...
host: 84.23.53.28:8080
info:
  contact: {}
  description: This is a server of YULA server.
  title: YULA project API
  version: "1.0"
paths:
...
  /order/add:
    post:
      consumes:
      - application/json
      description: add product in basket
      parameters:
      - description: order data for adding
        in: body
        name: preOrder
        required: true
        schema:
          $ref: '#/definitions/internal_models.PreOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.OrderResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: add order to basket
      tags:
      - order
  /order/buy_full_basket:
    patch:
      consumes:
      - application/json
      description: buy all orders from basket
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: buy all orders from basket
      tags:
      - order
  /order/delete/:
    delete:
      consumes:
      - application/json
      description: |-
        delete order for owner using user id from cookies\jwt.
        This totally removed order. Recovery will be impossible
      parameters:
      - description: order id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: delete order
      tags:
      - order
  /order/get_basket:
    get:
      consumes:
      - application/json
      description: get basket of orders by user id from cookie\jwt token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.OrderListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get basket of orders
      tags:
      - order
  /order/update_count:
    patch:
      consumes:
      - application/json
      description: update order count using user id from cookie\jwt token
      parameters:
      - description: order data for updating use only id and count
        in: body
        name: orderChanges
        required: true
        schema:
          $ref: '#/definitions/internal_models.OrderChanges'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update order count
      tags:
      - order
  /order/update_status:
    patch:
      consumes:
      - application/json
      description: update order status using user id from cookie\jwt token
      parameters:
      - description: order data for updating use only id and status
        in: body
        name: orderChanges
        required: true
        schema:
          $ref: '#/definitions/internal_models.OrderChanges'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update order status
      tags:
      - order
...
  /product/add-to-fav:
    post:
      consumes:
      - application/json
      description: add product to favs using product id from query and user id form
        cookie
      parameters:
      - description: product id
        in: query
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.ProductListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: add product to favs
      tags:
      - favourite
...
  /product/remove-from-fav:
    delete:
      consumes:
      - application/json
      description: delete product from favs using product id from query and user id
        form cookie
      parameters:
      - description: product id
        in: query
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.ProductListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: delete product from favs
      tags:
      - favourite
...
  /profile/profile/favourites:
    get:
      consumes:
      - application/json
      description: get user favourites by user id from cookie\jwt token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.ProductListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get user favourites
      tags:
      - favourite
...
schemes:
- http
swagger: "2.0"